Redistribute util.py (and its tests) to separate directory
