+++ sentence_transformers/util/decorators.py
@@ -1,0 +1,28 @@
+from __future__ import annotations
+
+import functools
+import logging
+
+logger = logging.getLogger(__name__)
+
+# TODO: decorator.py or decorators.py?
+
+
+def save_to_hub_args_decorator(func):
+    @functools.wraps(func)
+    def wrapper(self, *args, **kwargs):
+        # If repo_id not already set, use repo_name
+        repo_name = kwargs.pop("repo_name", None)
+        if repo_name and "repo_id" not in kwargs:
+            logger.warning(
+                "Providing a `repo_name` keyword argument to `save_to_hub` is deprecated, please use `repo_id` instead."
+            )
+            kwargs["repo_id"] = repo_name
+
+        # If positional args are used, adjust for the new "token" keyword argument
+        if len(args) >= 2:
+            args = (*args[:2], None, *args[2:])
+
+        return func(self, *args, **kwargs)
+
+    return wrapper
