+++ tests/util/test_tensor.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import numpy as np
+import torch
+
+from sentence_transformers import util
+
+
+def test_normalize_embeddings() -> None:
+    """Tests the correct computation of util.normalize_embeddings"""
+    embedding_size = 100
+    a = torch.tensor(np.random.randn(50, embedding_size))
+    a_norm = util.normalize_embeddings(a)
+
+    for embedding in a_norm:
+        assert len(embedding) == embedding_size
+        emb_norm = torch.norm(embedding)
+        assert abs(emb_norm.item() - 1) < 0.0001
